Adding failing test for annotated POST payloads.
If the POST payload has any kind of annotation, it is not picked up
by the doclet. This commit adds a regression for this fact.

We discovered this behaviour when trying to annotate a payload with
`@javax.validation.constraints.NotNull`. I did not want to introduce
a new test dependency, so I tried it with other annotations. It does
behave the same way with `@java.lang.Deprecated` as well.