Extract type/field/method to translation class and resolve name clashes.
Conflict resolving works for JAXB annotated type names, since the other
naming is based on the qualified type name, which should not clash.
The conflict is resolved by adding "_" characters at the end until it
is no longer conflicting.

Potentially this will resolve "conflicts" between classes not related
to each other, so a resource might have a documentation with class name
ending with "_", with similar class names.